ensure change event is fired on update
Some browsers (IE) do not ensure that the change event is fired when an
input element's value is updated via javascript. See:

https://stackoverflow.com/questions/44915351/onchange-event-and-programmatic-input-value-change-conflict-on-ie

In order to make cleave.js work in these browsers, we must manually
dispatch the change event so that IE picks up the newly formatted value
on the input element.